RC1.0

+++ app/src/main/java/com/progdist/egm/proyectopdist/domain/EditItem.kt
@@ -10,8 +10,8 @@ class EditItem {
                                 nameId: String,
                                 code: String,
                                 name: String,
-                                sale: Int,
-                                purchase: Int,
+                                sale: Float,
+                                purchase: Float,
                                 stock: Int,
                                 idCategory: Int,
                                 idSupplier: Int): Resource<Any> =
