+++ app/src/main/java/com/progdist/egm/proyectopdist/ui/home/owner/inventory/viewmodels/ItemSummaryViewModel.kt
@@ -61,8 +61,8 @@ class ItemSummaryViewModel(
         whereId: String,
         code: String,
         name: String,
-        sale: Int,
-        purchase: Int,
+        sale: Float,
+        purchase: Float,
         stock: Int,
         idCategory: Int,
         idSupplier: Int
